feat(week4): Fix Merge Error
