+++ app/src/main/java/org/sopt/doeuijin/EuijinApplication.kt
@@ -2,7 +2,6 @@ package org.sopt.doeuijin
 
 import android.app.Application
 import org.sopt.common.api.ApiFactory
-import org.sopt.common.api.ApiFactory2
 import org.sopt.doeuijin.container.SharedPreferenceContainer
 import org.sopt.doeuijin.container.SharedPreferenceContainer.sharedPreferencesInstance
 
@@ -12,7 +11,7 @@ class EuijinApplication : Application() {
         euijinApplicationInstance = this
         sharedPreferencesInstance = SharedPreferenceContainer.getAppPreferences(this)
         ApiFactory.retrofit = ApiFactory.getRetrofit(BuildConfig.AUTH_BASE_URL)
-        ApiFactory2.reqresRetrofit = ApiFactory2.getRetrofit(REQRES_BASE_URL)
+        ApiFactory.reqresRetrofit = ApiFactory.getRetrofit(REQRES_BASE_URL)
     }
 
     companion object {
