+++ data/user/src/main/java/org/sopt/doeuijin/data/user/api/ReqresServicePool.kt
@@ -1,8 +1,8 @@
 package org.sopt.doeuijin.data.user.api
 
-import org.sopt.common.api.ApiFactory2
+import org.sopt.common.api.ApiFactory
 import org.sopt.doeuijin.data.user.api.service.ReqresService
 
 object ReqresServicePool {
-    val authService = ApiFactory2.create<ReqresService>()
+    val authService = ApiFactory.createReqres<ReqresService>()
 }
